+++ src/grids/cvt/cvt.h
@@ -36,8 +36,8 @@ class CVT : public Grid
         // NOTE: we dont have support for changing "init","sample" and their
         // corresponding strings. 
         int init, sample;
-        size_t sample_num, it_max; 
-        size_t writeFreq;
+        unsigned int sample_num, it_max; 
+        unsigned int writeFreq;
 
     public:
         /*******************
@@ -57,8 +57,8 @@ class CVT : public Grid
         // the number of dimensions for the CVT. Dimension allows us to generate
         // a CVT in lower dimensions than the nodes (e.g., 3D node cloud with a
         // cvt on each xy plane) 
-        CVT (size_t nb_nodes, size_t dimension, size_t nb_locked=0, Density* density_function=NULL, size_t num_samples=10000, size_t max_num_iters=1000, size_t write_frequency=20, size_t sample_batch_size=1000); 
-        CVT (std::vector<NodeType>& nodes, size_t dimension, size_t nb_locked=0, Density* density_function=NULL, size_t num_samples=10000, size_t max_num_iters=1000, size_t write_frequency=20, size_t sample_batch_size=1000); 
+        CVT (unsigned int nb_nodes, unsigned int dimension, unsigned int nb_locked=0, Density* density_function=NULL, unsigned int num_samples=10000, unsigned int max_num_iters=1000, unsigned int write_frequency=20, unsigned int sample_batch_size=1000); 
+        CVT (std::vector<NodeType>& nodes, unsigned int dimension, unsigned int nb_locked=0, Density* density_function=NULL, unsigned int num_samples=10000, unsigned int max_num_iters=1000, unsigned int write_frequency=20, unsigned int sample_batch_size=1000); 
 
         virtual ~CVT() {
 #if USE_KDTREE
@@ -77,7 +77,7 @@ class CVT : public Grid
         virtual void displaceBoundaryNodes(int dim_num, int nb_bnd_nodes, double r_computed[], double r_updated[]) {;}
 
         void syncCVTandGrid() {
-            for (size_t i = 0; i < nb_bnd; i++) {
+            for (unsigned int i = 0; i < nb_bnd; i++) {
                 Vec3 nd;
                 for (int j = 0; j < dim_num; j++) {
                     nd[j] = generators[i*dim_num + j];
@@ -87,7 +87,7 @@ class CVT : public Grid
                 // TODO: boundary normals
                 //            this->getBoundaryNormal(i) = computeBoundaryNormal(bndry_pts[i]);
             }
-            for (size_t i = nb_bnd; i < nb_nodes; i++) {
+            for (unsigned int i = nb_bnd; i < nb_nodes; i++) {
                 Vec3 nd;
                 for (int j = 0; j < dim_num; j++) {
                     nd[j] = generators[i*dim_num + j];
