+++ src/grids/cvt/ellipsoid_cvt.cpp
@@ -580,13 +580,13 @@ void EllipsoidCVT::fillBoundaryPoints(int dim_num, int nb_nodes, int *seed, doub
 {
     std::vector<double> intg;
 
-    size_t high_nb_pts = 200;
+    unsigned int high_nb_pts = 200;
     double bnd_intg = computeBoundaryIntegral(*rho, high_nb_pts, intg);
     double dom_intg = computeDomainIntegral(high_nb_pts, *rho);
 
     // total nb points used to compute Voronoi mesh. 
     // Only (nb_interior_pts-nb_bnd) will be able to move freely
-    size_t tot_nb_pts = nb_nodes;
+    unsigned int tot_nb_pts = nb_nodes;
     // number of boundary points, automatically calculated
     printf("tot_nb_pts= %d\n", tot_nb_pts);
     printf("domain integral = %f\n", dom_intg);
@@ -605,8 +605,8 @@ void EllipsoidCVT::fillBoundaryPoints(int dim_num, int nb_nodes, int *seed, doub
 
 #if 0
     //EFB052611: This is the ellipse
-    size_t nb_computed_bnd = bnd_intg * sqrt(tot_nb_pts / dom_intg);
-    size_t nb_bnd_1 = 1. + 16. * tot_nb_pts * dom_intg / (bnd_intg * bnd_intg);
+    unsigned int nb_computed_bnd = bnd_intg * sqrt(tot_nb_pts / dom_intg);
+    unsigned int nb_bnd_1 = 1. + 16. * tot_nb_pts * dom_intg / (bnd_intg * bnd_intg);
     nb_bnd_1 = -bnd_intg * bnd_intg / (4. * dom_intg) * (1. - sqrt(nb_bnd_1));
     nb_computed_bnd = nb_bnd_1; // more accurate formula
     printf("calculated nb boundary pts: %d\n", nb_computed_bnd);
@@ -640,7 +640,7 @@ void EllipsoidCVT::fillBoundaryPoints(int dim_num, int nb_nodes, int *seed, doub
 //
 // GORDONs comment: Ideally, I should be integrating with respect to theta for 
 // more accuracy.
-double EllipsoidCVT::computeBoundaryIntegral(Density& rho, size_t npts, vector<double>& intg)
+double EllipsoidCVT::computeBoundaryIntegral(Density& rho, unsigned int npts, vector<double>& intg)
 {
     intg.resize(npts);
 #if 0
@@ -687,7 +687,7 @@ double EllipsoidCVT::computeBoundaryIntegral(Density& rho, size_t npts, vector<d
 
 //----------------------------------------------------------------------
 
-double EllipsoidCVT::computeDomainIntegral(size_t npts, Density& rho) {
+double EllipsoidCVT::computeDomainIntegral(unsigned int npts, Density& rho) {
     double major = axis_major; 
     double minor = axis_minor;
     double midax = axis_midax;
